Endre URLen til å være /barnebidrag/tjenester
