+++ .nais/prod.json
@@ -2,5 +2,5 @@
   "SERVER_URL": "http://bidrag-bidragskalkulator-api",
   "ENVIRONMENT": "prod",
   "UMAMI_WEBSITE_ID": "55af67ac-c5df-440e-80d6-518577e6de70",
-  "INGRESS": "https://www.nav.no/barnebidrag/kalkulator"
+  "INGRESS": "https://www.nav.no/barnebidrag/tjenester"
 }
