+++ app/routes.ts
@@ -1,6 +1,7 @@
 import { index, route, type RouteConfig } from "@react-router/dev/routes";
 export default [
   index("routes/index.tsx"),
+  route("/kalkulator", "routes/kalkulator.tsx"),
 
   route("/api/internal/isalive", "routes/api/internal/isalive.ts"),
   route("/api/internal/isready", "routes/api/internal/isready.ts"),
