+++ playwright.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
   globalSetup: "./e2e/global-setup.ts",
   globalTeardown: "./e2e/global-teardown.ts",
   use: {
-    baseURL: "http://localhost:5173/barnebidrag/kalkulator",
+    baseURL: "http://localhost:5173/barnebidrag/tjenester",
     trace: "on-first-retry",
   },
   projects: [
