+++ vite.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";
 
 export default defineConfig({
-  base: "/barnebidrag/kalkulator/",
+  base: "/barnebidrag/tjenester/",
   plugins: [
     tailwindcss(),
     reactRouterDevTools(),
